backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Blockchain
web3==6.11.1
eth-account==0.8.0
eth-typing==3.5.1
eth-utils==2.3.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Monitoring
prometheus-client==0.19.0
grafana-api==1.0.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pydantic==2.5.2
python-dateutil==2.8.2

# Security
cryptography==41.0.5
pyjwt==2.8.0 